Implement contact list and follow request features with authentication
- Added contact list functionality for users to view their contacts.
- Implemented full follow request flow: send, accept, decline, and reject requests.
- Integrated authentication with JWT for securing API endpoints.
- Updated routes to require authentication for contact and follow request actions.
- Enhanced user management and contact verification functionality.

+++ model/chat.go
@@ -10,5 +10,5 @@ type Chat struct {
 
 type ContactList struct {
 	Username     string `json:"username"`
-	LastActivity int64  `json:"last_activity"`
+	LastActivity float64  `json:"last_activity"`
 }
